Parcial implementation of new detectMapGestures
